requirements.txt
flask
load_dotenv
flask-sqlalchemy
psycopg2-binary